refactor: move image.lua to utils/base64.lua
