+++ lua/jet/core/kernel/init.lua
@@ -419,7 +419,7 @@ function Kernel:save_image(content, mime, name)
 		return false
 	end
 
-	return require("jet.core.image").base64_to_file(content, path)
+	return require("lua.jet.core.utils.base64").base64_to_file(content, path)
 end
 
 ---@private
